Serde xml deserialize prototype + refactor
Implemented a prototype of deserializing a basic imagedata xml file
using the serde facilities in quick_xml. This may not be the end goal,
but it is vastly simpler than using the quick_xml reader.

Added an Error for model operations. This is now required since we may
be loading more data using functions defined directly on the model. This
is because files in "parallel" xml formats are loaded lazily.

Defined DataArray model type to represent the (String, Attribute) pair
as a named attribute. This naming is consistent with the xml format
docs.

Defined VertexNumbers type to generalize over the way vertex numbers are
stored in XML vs Legacy vtk formats.

Defined the Coordinates type to contain x, y and z coordinates, the way
it is done in the XML format.

Defined Piece and PieceData types to allow DataSets to contain more than
one piece, which is allowed in XML formats. The reason Piece is distinct
from PieceData is that parallel xml formats allow lazy loading, so
Piece itself may represent only a URI to the file containing the actual
data, whereas PieceData is the data itself.

DataSet is simplified to contain primarily just a vec of pieces.

Fixed tests with model updates.