Minor refactor + renaming
Remove unused code and rename DataType -> ScalarType as well as the
Scalar variants on IOBuffer and the new ScalarType to correspond closer
to Rust types instead of VTK names. This makes code easier to read
albeit a bit less familiar to those who are used to the C/C++ primitive
types and Legacy VTK naming conventions.