Updated docs + small refactor
Added a CHANGELOG document for recording changes in the new 0.4 version.
Updated readme with more accurate and up-to-date information.

Removed ScalarArray from existence and replaced it with a simple
IOBuffer because Coordinates don't need to be named.

Added iter and into_vec functions for IOBuffer for better compatibility
with previous version of IOBuffer.