Simplify writing code
Added buffered writing and avoid in-memory writes (which can use up a
lot of memory).
Writing trait is also made more generic in the process.
Fix up export commands to include the correct ByteOrder tag. This isn't
currently use since byte order is determined entirely at compile time,
but it is included for future proofing: if this tag is used later
on, we make sure it is correct.