Deprecate import_{le,be} + fix byteorder on parse
Byte order was always set to BigEndian even if the file was parse as
little endian. Although this doesn't have any impact on the parsing
itself, it is confusing to get a Vtk struct back with a different
endianness. Thus it is considered a bug.

import_{le,be} functions are deprecated in favour of
import_legacy_{le,be} to be consistent with the naming of the other
functions.