Added load_piece_data function + update year
Load_piece_data takes the piece by reference thus can be called multiple
times on the same piece. This is convenient when loading shouldn't
consume the piece, although most likely the user would want to preload
everything, so this funciton should be used as a getter.

Also updated the year in the license.