Update legacy parser to work with offsets
Added a parser for OFFSETS and CONNECTIVITY tags within legacy vtk files
that mimic the way the xml file work.

There is a new version of legacy vtk files (seemingly v5 and up) that
specify cells via OFFSETS and CONNECTIVITY tags. This parser adds
support for that.

See issue #21 for exmaples.

This commit addresses the parsing of the legacy files in that issue.