Unified versions between xml and legacy file types
This commit switches model::Version to be an enum with legacy and xml
variants to indicate which type of file the data came from and what
version it was.

This does not preclude the model::Vtk struct from being serialized
either as legacy or xml formats. That functionality remains flexible.

Fixes #12