Refactor to support latest quick-xml version
This commit sadly breaks support for VTK files with binary AppendedData
blobs.
The code needed to support this is currently not functional and
depends on the completion of this feature in quick-xml
(tracked in https://github.com/tafia/quick-xml/issues/623)
All binary related code is now featured gated, which is a regression
from previous version of vtkio, but will be re-enabled in the future.